First GraphAPI prototype

+++ src/main/java/com/towels/graphofcontent/dto/EdgeDTO.java
@@ -2,7 +2,7 @@ package com.towels.graphofcontent.dto;
 
 
 public class EdgeDTO {
-	public int id;
-	public int source;
-	public int target;
+	public long id;
+	public long source;
+	public long target;
 }
